BUGFIX More resilient parsing against UTF8 errors (just happens on ernie server it seems)
git-svn-id: http://svn.silverstripe.com/projects/ss2api/trunk@108701 467b73ca-7a2a-4603-9d3b-597d59a354a9

+++ htdocs/search/mysite/code/tools/SSAPIGotApiImporter.php
@@ -53,7 +53,16 @@ class SSAPIGotApiImporter {
 	function run() {
 		$ids = array();
 		
-		$xml = simplexml_load_file($this->filePath);
+		libxml_use_internal_errors(true);
+		$xmlStr = file_get_contents($this->filePath);
+		// file should be UTF8, but who knows ... stop libxml from complaining
+		$xmlStr = utf8_encode($xmlStr);
+		$xml = simplexml_load_string($xmlStr);
+		if(!$xml) {
+			var_dump(libxml_get_errors());
+			return false;
+		}
+		
 		if($xml->page) {
 			// Clear out existing entries
 			DB::query(sprintf('DELETE FROM "SSAPIProperty" WHERE "VersionString" = \'%s\'', $this->version));
